chore(fias): rename files in Fias
