+++ packages/react-ui/components/Fias/Form/__tests__/FiasComboBox-test.tsx
@@ -4,7 +4,7 @@ import { mount } from 'enzyme';
 import { FiasComboBox } from '../FiasComboBox';
 import { delay } from '../../../../lib/utils';
 import { Menu } from '../../../Menu';
-import { FiasAddress } from '../../models/Address';
+import { FiasAddress } from '../../models/FiasAddress';
 
 function searchFactory<T>(promise: Promise<T>): [jest.Mock<Promise<T>>, Promise<T>] {
   let searchCalled: () => void;
