+++ packages/react-ui/components/Fias/index.ts
@@ -1,4 +1,4 @@
 export * from './Fias';
 export * from './FiasSearch/FiasSearch';
 export * from './api/FiasAPI';
-export * from './models/Address';
+export * from './models/FiasAddress';
