+++ packages/react-ui/components/Fias/models/FiasAddress.ts
@@ -18,7 +18,7 @@ import {
   FiasAPIProvider,
 } from '../types';
 
-import { FiasAddressElement } from './AddressElement';
+import { FiasAddressElement } from './FiasAddressElement';
 import { FiasData } from './FiasData';
 
 export interface FiasAddressOptions {
